Build POC gRPC server
