+++ .gitignore
@@ -17,3 +17,4 @@ ruruku-*.yaml
 build/release
 build/deployKey
 **/rice-box.go
+**/*.pb.go
